Added stamina mechanic, refined enemies
Stamina bar is now in game, when player sprints it gets lower, until it gets to 0 and player can't sprint. It will charge back up when player stops sprinting (at lower pace than it gets used).
Enemies are now quicker than the player, forcing the player to use the sprint, which is faster than them